Move SEO and SiteMetadata components to Layout/
