+++ src/components/Show/index.tsx
@@ -5,7 +5,7 @@ import FaChevronLeft from 'react-icons/lib/fa/chevron-left';
 import FaExternalLink from 'react-icons/lib/fa/external-link';
 import { OutboundLink } from 'gatsby-plugin-google-analytics';
 
-import { SEOTypes } from '../SEO';
+import { SEOTypes } from '../Layout/SEO';
 import Layout from '../Layout/Layout';
 import Page from '../Layout/Page';
 import { WORK_URL } from '../../constants/urls';
