+++ src/pages/404.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'gatsby';
 
 import Layout from '../components/Layout/Layout';
-import { SEOTypes } from '../components/SEO';
+import { SEOTypes } from '../components/Layout/SEO';
 import Page from '../components/Layout/Page';
 import Button, { ButtonThemes, ButtonSizes } from '../components/common/Button';
