+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import cx from 'classnames';
 import { Link, graphql } from 'gatsby';
 import Img from 'gatsby-image';
 
-import { SEOTypes } from '../components/SEO';
+import { SEOTypes } from '../components/Layout/SEO';
 import Layout from '../components/Layout/Layout';
 import BlogIndexPost from '../components/BlogIndexPost';
 import { HOME_URL } from '../constants/urls';
