+++ src/pages/work.tsx
@@ -5,7 +5,7 @@ import Link from 'gatsby-link';
 import { OutboundLink } from 'gatsby-plugin-google-analytics';
 import FaExternalLink from 'react-icons/lib/fa/external-link';
 
-import { SEOTypes } from '../components/SEO';
+import { SEOTypes } from '../components/Layout/SEO';
 import Layout from '../components/Layout/Layout';
 import { WORK_URL, HOME_URL } from '../constants/urls';
 import { getDate } from '../utils/helpers';
